Reorganize the data
The data that used to reside in "Archive" was reviewed; obsolete layers
were removed, the rest was moved to vidnoe.gpkg. CSV and GeoJSON files
were removed because they can be recreated from the corresponding GPKG
tables at any time. Images were removed as unnecessary.
vidnoe_geoalert_vs_bti was removed since that data doesn't belong to
the benchmark. Address points remained in a separate gpkg to avoid
cluttering the main one since those addresses are not frequently used
anymore but are not entirely obslete either, so should be kept.